% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_find.R
\name{hk_find}
\alias{hk_find}
\title{Function for finding a group of relatively invariant OTUs}
\usage{
hk_find(OTU_table, min_avg_counts = 5)
}
\arguments{
\item{OTU_table}{pool for OTUs which will be tested for differential abundance in form of OTUs x Samples and rownames of OTU_table are OTU ids}

\item{min_avg_counts}{minimum average counts required for riOTUs}
}
\description{
This function allows you to find a group of relatively invariant OTUs and calculate size factors for normalization.
}
\keyword{OTUs}
\keyword{factors}
\keyword{invariant}
\keyword{normalization}
\keyword{relatively}
\keyword{size}
