% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hknorm.R
\name{hknorm}
\alias{hknorm}
\title{A normalization function}
\usage{
hknorm(OTU_table, size_factor)
}
\arguments{
\item{OTU_table}{pool for OTUs which will be tested for differential abundance in form of OTUs x Samples}

\item{size_factor}{size factors of all samples for normalization use}
}
\description{
This function allows you to calculate normalized OTU table.
}
\keyword{OTUs}
\keyword{RiOTUs}
\keyword{a}
\keyword{factors}
\keyword{group}
\keyword{invariant}
\keyword{normalization}
\keyword{of}
\keyword{relatively}
\keyword{size}
