% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp_scoretest.R
\name{overdisp_scoretest}
\alias{overdisp_scoretest}
\title{A overdispersion test function}
\usage{
overdisp_scoretest(OTU_table, class, log_sizefactor)
}
\arguments{
\item{OTU_table}{pool for OTUs which will be tested for differential abundance in form of OTUs x Samples}

\item{class}{labels for each sample}

\item{log_sizefactor}{log of size factors}
}
\description{
This function allows you to test for over-dispersion in zero-inflated count regression model
}
\keyword{overdispersion}
\keyword{score}
\keyword{test}
