% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RioNorm2_test.R
\name{RioNorm2_test}
\alias{RioNorm2_test}
\title{Function for RioNorm2 differential abundance test}
\usage{
RioNorm2_test(OTU_table, class, FDR = 1)
}
\arguments{
\item{OTU_table}{pool for OTUs which will be tested for differential abundance in form of OTUs x Samples.}

\item{class}{dummy variables for class information.}

\item{FDR}{level for FDR control}
}
\description{
This function allows you to apply RioNorm2 framework and detect differentially abundant OTUs.
}
\keyword{Differential}
\keyword{RioNorm2}
\keyword{abundance}
\keyword{analysis}
