% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB_test.R
\name{ZINB_test}
\alias{ZINB_test}
\title{A differential abundance test function}
\usage{
ZINB_test(disp_OTU, class, log_sizefactor)
}
\arguments{
\item{disp_OTU}{overdispersed OTU table in form of OTUs x samples}

\item{class}{labels for each sample}

\item{log_sizefactor}{log of size factors}
}
\description{
This function is for differential abundance test using zero-inflated Negative Binomial model
}
\keyword{Benjamini-Hochberg}
\keyword{abundance}
\keyword{differential}
\keyword{over-dispersion}
\keyword{test}
