% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP_test.R
\name{ZIP_test}
\alias{ZIP_test}
\title{A differential abundance test function}
\usage{
ZIP_test(nondisp_OTU, class, log_sizefactor)
}
\arguments{
\item{nondisp_OTU}{non-overdispersed OTU table in form of OTUs x samples}

\item{class}{labels for each sample}

\item{log_sizefactor}{log of size factors}
}
\description{
This function is for differential abundance test using zero-inflated Poisson model
}
\keyword{Benjamini-Hochberg}
\keyword{abundance}
\keyword{differential}
\keyword{non-overdispersion}
\keyword{test}
